+++ README.md
@@ -13,3 +13,4 @@
 - chore: 빌드 작업 또는 설정 변경
 - perf: 성능 개선을 위한 코드 변경
 - ci: CI 설정 파일 및 스크립트 변경
+- design : CSS 파일 변경
