+++ src/main/java/com/kosa/moimeasy/membership/controller/PayController.java
@@ -1,0 +1,4 @@
+package com.kosa.moimeasy.membership.controller;
+
+public class PayController {
+}
