+++ src/main/java/com/kosa/moimeasy/membership/dto/PayDTO.java
@@ -1,0 +1,4 @@
+package com.kosa.moimeasy.membership.dto;
+
+public class PayDTO {
+}
