+++ src/main/java/com/kosa/moimeasy/membership/entity/Pay.java
@@ -1,0 +1,4 @@
+package com.kosa.moimeasy.membership.entity;
+
+public class Pay {
+}
