+++ src/main/java/com/kosa/moimeasy/membership/repository/PayRepository.java
@@ -1,0 +1,4 @@
+package com.kosa.moimeasy.membership.repository;
+
+public interface PayRepository {
+}
