+++ src/main/java/com/kosa/moimeasy/membership/service/PayService.java
@@ -1,0 +1,4 @@
+package com.kosa.moimeasy.membership.service;
+
+public interface PayService {
+}
