--- src/main/java/com/kosa/moimeasy/pay/controller/PayController.java
@@ -1,4 +1,0 @@
-package com.kosa.moimeasy.pay.controller;
-
-public class PayController {
-}
