--- src/main/java/com/kosa/moimeasy/pay/dto/PayDTO.java
@@ -1,4 +1,0 @@
-package com.kosa.moimeasy.pay.dto;
-
-public class PayDTO {
-}
