--- src/main/java/com/kosa/moimeasy/pay/entity/Pay.java
@@ -1,4 +1,0 @@
-package com.kosa.moimeasy.pay.entity;
-
-public class Pay {
-}
