--- src/main/java/com/kosa/moimeasy/pay/repository/PayRepository.java
@@ -1,4 +1,0 @@
-package com.kosa.moimeasy.pay.repository;
-
-public interface PayRepository {
-}
