--- src/main/java/com/kosa/moimeasy/pay/service/PayService.java
@@ -1,4 +1,0 @@
-package com.kosa.moimeasy.pay.service;
-
-public interface PayService {
-}
